requirements.txt
facebook-business>=19.0.0
python-telegram-bot[job-queue]>=22.0
python-dotenv>=1.0.0
